+++ src/Events/CollectionHasBeenCleared.php
@@ -9,14 +9,10 @@ class CollectionHasBeenCleared
 {
     use SerializesModels;
 
-    /**
-     * @var \Spatie\MediaLibrary\HasMedia\Interfaces\HasMedia
-     */
+    /** @var \Spatie\MediaLibrary\HasMedia\Interfaces\HasMedia */
     public $model;
 
-    /**
-     * @var string
-     */
+    /** @var string */
     public $collectionName;
 
     public function __construct(HasMedia $model, string $collectionName)