+++ src/Events/ConversionHasBeenCompleted.php
@@ -10,19 +10,13 @@ class ConversionHasBeenCompleted
 {
     use SerializesModels;
 
-    /**
-     * @var \Spatie\MediaLibrary\Media
-     */
+    /** @var \Spatie\MediaLibrary\Media */
     public $media;
 
-    /**
-     * @var \Spatie\MediaLibrary\Conversion\Conversion
-     */
+    /** @var \Spatie\MediaLibrary\Conversion\Conversion */
     public $conversion;
 
     /**
-     * ConversionHasFinishedEvent constructor.
-     *
      * @param \Spatie\MediaLibrary\Media                 $media
      * @param \Spatie\MediaLibrary\Conversion\Conversion $conversion
      */