+++ tests/FileAdder/IntegrationTest.php
@@ -61,7 +61,9 @@ class IntegrationTest extends TestCase
     {
         $collectionName = 'images';
 
-        $media = $this->testModel->addMedia($this->getTestJpg())->toCollection($collectionName);
+        $media = $this->testModel
+            ->addMedia($this->getTestJpg())
+            ->toMediaLibrary($collectionName);
 
         $this->assertEquals($collectionName, $media->collection_name);
     }
@@ -84,7 +86,9 @@ class IntegrationTest extends TestCase
     {
         $testFile = $this->getTestJpg();
 
-        $media = $this->testModel->copyMedia($testFile)->toCollection('images');
+        $media = $this->testModel
+            ->copyMedia($testFile)
+            ->toMediaLibrary('images');
 
         $this->assertFileExists($testFile);
         $this->assertFileExists($this->getMediaDirectory($media->id.'/'.$media->file_name));
@@ -93,7 +97,9 @@ class IntegrationTest extends TestCase
     /** @test */
     public function it_can_handle_a_file_without_an_extension()
     {
-        $media = $this->testModel->addMedia($this->getTestFilesDirectory('test'))->toMediaLibrary();
+        $media = $this->testModel
+            ->addMedia($this->getTestFilesDirectory('test'))
+            ->toMediaLibrary();
 
         $this->assertEquals('test', $media->name);
 
@@ -107,7 +113,9 @@ class IntegrationTest extends TestCase
     /** @test */
     public function it_can_handle_an_image_file_without_an_extension()
     {
-        $media = $this->testModel->addMedia($this->getTestFilesDirectory('image'))->toMediaLibrary();
+        $media = $this->testModel
+            ->addMedia($this->getTestFilesDirectory('image'))
+            ->toMediaLibrary();
 
         $this->assertEquals(Media::TYPE_IMAGE, $media->type);
     }
@@ -115,7 +123,9 @@ class IntegrationTest extends TestCase
     /** @test */
     public function it_can_handle_a_non_image_and_non_pdf_file()
     {
-        $media = $this->testModel->addMedia($this->getTestFilesDirectory('test.txt'))->toMediaLibrary();
+        $media = $this->testModel
+            ->addMedia($this->getTestFilesDirectory('test.txt'))
+            ->toMediaLibrary();
 
         $this->assertEquals('test', $media->name);
 
@@ -136,7 +146,10 @@ class IntegrationTest extends TestCase
             filesize($this->getTestFilesDirectory('test.jpg'))
         );
 
-        $media = $this->testModel->addMedia($uploadedFile)->toMediaLibrary();
+        $media = $this->testModel
+            ->addMedia($uploadedFile)
+            ->toMediaLibrary();
+
         $this->assertEquals('alternativename', $media->name);
         $this->assertFileExists($this->getMediaDirectory($media->id.'/'.$media->file_name));
     }
@@ -145,7 +158,10 @@ class IntegrationTest extends TestCase
     public function it_can_add_an_upload_to_the_medialibrary_from_the_current_request()
     {
         $this->app['router']->get('/upload', function () {
-            $media = $this->testModel->addMediaFromRequest('file')->toMediaLibrary();
+            $media = $this->testModel
+                ->addMediaFromRequest('file')
+                ->toMediaLibrary();
+
             $this->assertEquals('alternativename', $media->name);
             $this->assertFileExists($this->getMediaDirectory($media->id.'/'.$media->file_name));
         });
@@ -169,7 +185,9 @@ class IntegrationTest extends TestCase
             $exceptionWasThrown = false;
 
             try {
-                $this->testModel->addMediaFromRequest('non existing key')->toMediaLibrary();
+                $this->testModel
+                    ->addMediaFromRequest('non existing key')
+                    ->toMediaLibrary();
             } catch (RequestDoesNotHaveFile $exception) {
                 $exceptionWasThrown = true;
             }
@@ -316,6 +334,6 @@ class IntegrationTest extends TestCase
 
         $this->testModel
             ->addMedia($this->getTestJpg())
-            ->toMediaLibraryOnDisk('images', 'non-existing-disk');
+            ->toMediaLibrary('images', 'non-existing-disk');
     }
 }