+++ tests/HasMediaTrait/HasMediaTest.php
@@ -23,7 +23,7 @@ class HasMediaTest extends TestCase
     /** @test */
     public function it_returns_true_for_if_any_collection_is_not_empty()
     {
-        $this->testModel->addMedia($this->getTestJpg())->toCollection('images');
+        $this->testModel->addMedia($this->getTestJpg())->toMediaLibrary('images');
 
         $this->assertTrue($this->testModel->hasMedia('images'));
     }
@@ -37,7 +37,7 @@ class HasMediaTest extends TestCase
     /** @test */
     public function it_returns_true_for_a_non_empty_named_collection()
     {
-        $this->testModel->addMedia($this->getTestJpg())->toCollection('images');
+        $this->testModel->addMedia($this->getTestJpg())->toMediaLibrary('images');
 
         $this->assertTrue($this->testModel->hasMedia('images'));
         $this->assertFalse($this->testModel->hasMedia('downloads'));